requirements.txt
discord.py
colorama
numpy
requests
aiohttp
bs4
pymongo
selenium
pyPrivnote
gtts
dnspython
image
datetime
asyncio